+++ webpack.common.js
@@ -59,8 +59,4 @@ module.exports = {
     contentBase: './docs/dist',
     port: 8080,
   },
-
-  node: {
-    fs: 'empty',
-  }
 };
